Move my class in my packages
